删除账户前检测余额

+++ includes/action/wnd-delete-user.php
@@ -29,6 +29,16 @@ class Wnd_Delete_User extends Wnd_Action_Ajax {
 			throw new Exception(__('请确认操作', 'wnd'));
 		}
 
+		/**
+		 *@since 0.8.64
+		 *
+		 *删除用户权限检测过滤
+		 */
+		$wnd_can_delete_user = apply_filters('wnd_can_delete_user', ['status' => 1, 'msg' => ''], $user_id);
+		if (0 === $wnd_can_delete_user['status']) {
+			return $wnd_can_delete_user;
+		}
+
 		require_once ABSPATH . 'wp-admin/includes/user.php';
 		$action = wp_delete_user($user_id);
 		if ($action) {
